requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
fake-useragent==1.4.0
greenlet==3.0.1
idna==3.5
psycopg2==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
